Support if in if Remove java from ql code Add node enum types

+++ Joanna_Laurence/src/main/java/org/uva/sea/ql/QLExprEvaluate.java
@@ -70,7 +70,7 @@ public class QLExprEvaluate extends Traverse {
     private ASTNode makeTypeCompatible(ASTNode node, ASTNode with) {
         if(node instanceof Int && with instanceof Dec) {
             int intVal = ((Int)node).getValue();
-            return new Dec((double)intVal);
+            return new Dec(intVal);
         }
 
         return node;
